@@ -1,10 +1,13 @@
 
-import { Observable, ReplaySubject } from 'rxjs/Rx';
+import { Observable, ReplaySubject } from 'rxjs';
 
 import { assert, log, logIf, LogLevel } from './auxiliaries';
 import { clamp } from './gl-matrix-extensions';
 
-
+/**
+ * This internal interface is not intended for module export and should not be used as generic rendering/controller
+ * interface. The renderer interface should be used for rendering related tasks instead.
+ */
 export interface Controllable {
 
     /**
@@ -52,7 +55,8 @@ export class Controller {
      */
     protected _batchSize = 1;
     set batch(size: number) {
-        log(LogLevel.Dev, `(adaptive) batch multi-frame rendering is experimental for now`);
+        log(LogLevel.Warning, `(adaptive) batch multi-frame rendering is experimental for now`);
+        this._batchSize = Math.max(1, size);
     }
 
 
@@ -167,14 +171,14 @@ export class Controller {
          * actual first frame is triggered.
          */
         if (this._pendingRequest !== 0) {
-            logIf(this._debug, LogLevel.ModuleDev, `c request (ignored) | pending: '${this._pendingRequest}'`);
+            logIf(this._debug, LogLevel.Debug, `c request (ignored) | pending: '${this._pendingRequest}'`);
             return;
         }
         if (this._pause || !this._controllable) {
-            logIf(this._debug, LogLevel.ModuleDev, `c request (ignored) | pending: '${this._pendingRequest}'`);
+            logIf(this._debug, LogLevel.Debug, `c request (ignored) | pending: '${this._pendingRequest}'`);
             return;
         }
-        logIf(this._debug, LogLevel.ModuleDev, `c request           | intermediates: #${this._frameNumber}`);
+        logIf(this._debug, LogLevel.Debug, `c request           | intermediates: #${this._frameNumber}`);
 
         const dfnum = this._debugFrameNumber;
         const mfnum = this._multiFrameNumber;
@@ -199,7 +203,7 @@ export class Controller {
 
     protected reset(): boolean {
         const block = this._block || (this._frameNumber === 0 && this._pendingRequest);
-        logIf(this._debug, LogLevel.ModuleDev, `c update  ${block ? '(blocked) ' : '          '}| ` +
+        logIf(this._debug, LogLevel.Debug, `c update  ${block ? '(blocked) ' : '          '}| ` +
             `pending: '${this._pendingRequest}', intermediates: #${this._frameNumber}`);
 
         if (block) {
@@ -215,10 +219,10 @@ export class Controller {
      */
     protected cancel(): void {
         if (this._pendingRequest === 0) {
-            logIf(this._debug, LogLevel.ModuleDev, `c cancel  (ignored) |`);
+            logIf(this._debug, LogLevel.Debug, `c cancel  (ignored) |`);
             return;
         }
-        logIf(this._debug, LogLevel.ModuleDev, `c cancel            | pending: '${this._pendingRequest}'`);
+        logIf(this._debug, LogLevel.Debug, `c cancel            | pending: '${this._pendingRequest}'`);
 
         window.cancelAnimationFrame(this._pendingRequest);
         this._pendingRequest = 0;
@@ -248,7 +252,7 @@ export class Controller {
     }
 
     protected invokeUpdate(force: boolean = false): void {
-        logIf(this._debug, LogLevel.ModuleDev, `c invoke update     | ` +
+        logIf(this._debug, LogLevel.Debug, `c invoke update     | ` +
             `pending: '${this._pendingRequest}', mfnum: ${this._multiFrameNumber}`);
 
         this.unblock();
@@ -266,7 +270,7 @@ export class Controller {
      * Actual invocation of the controllable's prepare method
      */
     protected invokePrepare(): void {
-        logIf(this._debug, LogLevel.ModuleDev, `c invoke prepare    |`);
+        logIf(this._debug, LogLevel.Debug, `c invoke prepare    |`);
 
         this._frameNumber = 0;
 
@@ -296,7 +300,7 @@ export class Controller {
      */
     protected invokeFrame(): void {
         assert(!this._pause, `frames should not be invoked when paused`);
-        logIf(this._debug, LogLevel.ModuleDev, `c invoke frame      | pending: '${this._pendingRequest}'`);
+        logIf(this._debug, LogLevel.Debug, `c invoke frame      | pending: '${this._pendingRequest}'`);
 
         const dfnum = this._debugFrameNumber;
         const mfnum = this._multiFrameNumber;
@@ -322,11 +326,11 @@ export class Controller {
         }
 
         for (; this._frameNumber < batchEnd; ++this._frameNumber) {
-            logIf(this._debug, LogLevel.ModuleDev, `c -> frame          | frame: ${this._frameNumber}`);
+            logIf(this._debug, LogLevel.Debug, `c -> frame          | frame: ${this._frameNumber}`);
             (this._controllable as Controllable).frame(this._frameNumber);
             ++this._intermediateFrameCount;
         }
-        logIf(this._debug, LogLevel.ModuleDev, `c -> swap           |`);
+        logIf(this._debug, LogLevel.Debug, `c -> swap           |`);
 
         (this._controllable as Controllable).swap();
         this._multiTime[1] = performance.now();
@@ -374,7 +378,7 @@ export class Controller {
      */
     pause(): void {
         const ignore = this._pause;
-        logIf(this._debug, LogLevel.ModuleDev, `c pause   ${ignore ? '(ignored)' : ''}`);
+        logIf(this._debug, LogLevel.Debug, `c pause   ${ignore ? '(ignored)' : ''}`);
 
         if (this._pause) {
             return;
@@ -390,7 +394,7 @@ export class Controller {
      */
     unpause(): void {
         const ignore = !this._pause;
-        logIf(this._debug, LogLevel.ModuleDev, `c unpause ${ignore ? '(ignored)' : ''}`);
+        logIf(this._debug, LogLevel.Debug, `c unpause ${ignore ? '(ignored)' : ''}`);
 
         if (ignore) {
             return;
@@ -432,7 +436,7 @@ export class Controller {
      * triggering to multiple intermediate updates. The block updates mode can be exited using ``unblock```.
      */
     block(): void {
-        logIf(this._debug, LogLevel.ModuleDev, `c block   ${this._block ? '(ignored) ' : '          '}|`);
+        logIf(this._debug, LogLevel.Debug, `c block   ${this._block ? '(ignored) ' : '          '}|`);
 
         if (this._block) {
             return;
@@ -444,7 +448,7 @@ export class Controller {
      * Unblock updates. If there was at least one blocked update request, an immediate update is invoked.
      */
     unblock(): void {
-        logIf(this._debug, LogLevel.ModuleDev, `c unblock ${!this._block ? '(ignored) ' : '          '}` +
+        logIf(this._debug, LogLevel.Debug, `c unblock ${!this._block ? '(ignored) ' : '          '}` +
             `| blocked: #${this._blockedUpdates}`);
 
         if (!this._block) {
@@ -513,7 +517,7 @@ export class Controller {
         this._multiFrameNumber = value;
         this.multiFrameNumberNext();
 
-        logIf(value !== multiFrameNumber, LogLevel.ModuleDev,
+        logIf(value !== multiFrameNumber, LogLevel.Debug,
             `multi-frame number adjusted to ${value}, given ${multiFrameNumber}`);
 
         if (this.debugFrameNumber > this.multiFrameNumber) {
@@ -562,7 +566,7 @@ export class Controller {
         this._debugFrameNumber = value;
         this.debugFrameNumberNext();
 
-        logIf(value !== debugFrameNumber, LogLevel.ModuleDev,
+        logIf(value !== debugFrameNumber, LogLevel.Debug,
             `debug-frame number adjusted to ${value}, given ${debugFrameNumber}`);
 
         if (this._block) {
@@ -664,7 +668,6 @@ export class Controller {
     get multiFrameTime(): number {
         return this._frameNumber === 0 ? 0.0 : this._multiTime[1] - this._multiTime[0];
     }
-
     /**
      * The frames per second is based on the average number of a full intermediate frame request up to the current
      * frame number.
