@@ -13,15 +13,17 @@ const pug = require('pug');
 const copy = require('./copy.js');
 
 const websiteDir = './website';
-const distDir = './build';
+const buildDir = './build';
 
 const entries = ['index.pug'];
 
 const assets = [
-    [websiteDir, distDir, ['css/*.css', 'js/*.js', 'img/*.{svg,png}', 'fonts/*', '*.{svg,png,ico,xml,json}'], [], false],
-    ['./source/data/', distDir+'/data', ['*'], [], false],
-    ['./source/data/opensansr144', distDir+'/data/opensansr144', ['*'], [], false],
-    ['./source/data/verdana', distDir+'/data/verdana', ['*'], [], false]];
+    [websiteDir, buildDir, ['css/*.css', 'js/*.js', 'img/*.{svg,png}', 'fonts/*', '*.{svg,png,ico,xml,json}'], [], false],
+    ['./source', buildDir, ['data/{colorbrewer,smithwalt}.json'], [], false],
+    ['./source/data/', buildDir+'/data', ['*'], [], false],
+    ['./source/data/opensansr144', buildDir+'/data/opensansr144', ['*'], [], false],
+    ['./source/data/verdana', buildDir+'/data/verdana', ['*'], [], false]];
+    ['./node_modules/rxjs/bundles/', `${buildDir}/js`, ['rxjs.umd.min.js'], [], false]];
 
 
 var build_pending = false;
@@ -31,7 +33,7 @@ function build() {
 
     entries.forEach((entry) => {
         const src = path.join(websiteDir, entry);
-        const dst = path.join(distDir, path.basename(entry, path.extname(entry)) + '.html');
+        const dst = path.join(buildDir, path.basename(entry, path.extname(entry)) + '.html');
 
         if (!fs.existsSync(src)) {
             console.log('skipped:', entry);
