@@ -1,6 +1,7 @@
 
 import { assert } from './auxiliaries';
 
+import { Context } from './context';
 import { WEBGL1_EXTENSIONS, WEBGL2_EXTENSIONS } from './extensions';
 
 export class ExtensionsHash {
@@ -86,7 +87,7 @@ export class ExtensionsHash {
      * @param backend - WebGL backend: 'webgl1' or 'webgl2'.
      * @param supported - Array of supported extensions to be encoded.
      */
-    static encode(backend: string, supported: Array<string>): string {
+    static encode(backend: Context.BackendType, supported: Array<string>): string {
         const version = ExtensionsHash.LATEST_VERSION; // should always be set to the latest version
         const extensions = ExtensionsHash.EXTENSIONS_BY_VERSION.get(version) as Array<string>;
 
