@@ -12,17 +12,21 @@ import { Shader } from './shader';
  * uniform location retrieval. A program is intended to be used as follows:
  *
  * ```
- * this.program.initialize(
- *  [('screenaligned.vert', require('../shaders/screenaligned.vert'))],
- *  [('blit.frag', require('../shaders/blit.frag'))]);
+ * const vert = new Shader(this._context, gl.VERTEX_SHADER, 'ndcvertices.vert (blit)');
+ * vert.initialize(require('./shaders/ndcvertices.vert'));
+ * const frag = new Shader(this._context, gl.FRAGMENT_SHADER, 'blit.frag');
+ * frag.initialize(require('./shaders/blit.frag'));
  *
- * this.a_vertex = this.program.attribute('a_vertex');
- * const u_texture = this.program.uniform('u_texture');
+ * this._program = new Program(this._context, 'BlitProgram');
+ * this._program.initialize([vert, frag]);
  *
- * this.program.bind();
- * gl.uniform1i(u_texture, 0);
+ * this.aVertex = this._program.attribute('a_vertex');
+ * const uTexture = this._program.uniform('u_texture');
+ *
+ * this._program.bind();
+ * gl.uniform1i(uTexture, 0);
  * // ... draw
- * this.program.unbind();
+ * this._program.unbind();
  * ```
  */
 export class Program extends AbstractObject<WebGLProgram> implements Bindable {
@@ -45,7 +49,7 @@ export class Program extends AbstractObject<WebGLProgram> implements Bindable {
 
         for (const shader of shaders) {
             if (!shader.initialized) {
-                log(LogLevel.Dev, `shader '${shader.identifier}' not initialized.`);
+                log(LogLevel.Error, `shader '${shader.identifier}' not initialized.`);
                 continue;
             }
             gl.attachShader(this._object, shader.object);
@@ -67,7 +71,7 @@ export class Program extends AbstractObject<WebGLProgram> implements Bindable {
 
         if (!gl.getProgramParameter(this._object, gl.LINK_STATUS)) {
             const infoLog: string = gl.getProgramInfoLog(this._object);
-            log(LogLevel.Dev, `linking of program '${this._identifier}' failed: '${infoLog}'`);
+            log(LogLevel.Error, `linking of program '${this._identifier}' failed: '${infoLog}'`);
             return false;
         }
         return true;
@@ -114,8 +118,8 @@ export class Program extends AbstractObject<WebGLProgram> implements Bindable {
                     break;
             }
         }
-        logIf(numVertShaders < 1, LogLevel.Dev, `at least one vertex shader is expected`);
-        logIf(numFragShaders < 1, LogLevel.Dev, `at least one fragment shader is expected`);
+        logIf(numVertShaders < 1, LogLevel.Error, `at least one vertex shader is expected`);
+        logIf(numFragShaders < 1, LogLevel.Error, `at least one fragment shader is expected`);
         if (numVertShaders < 1 || numFragShaders < 1) {
             return undefined;
         }
