@@ -1,25 +1,27 @@
 
-module.exports = require('./webpack.config');
+module.exports = (env, options) => {
 
-module.exports.cache = false;
-module.exports.entry = {
-    '../lib/webgl-operate': ['externals.ts', 'require.ts', 'polyfill.ts', 'webgl-operate.ts']
-};
+    const config = require('./webpack.config');
 
-module.exports.module.rules[0].use = {
-    loader: 'ts-loader',
-    options: {
-        compilerOptions: {
-            declaration: true,
-            removeComments: false
-        }
-    }
-};
+    config.cache = false;
+    config.output.path = __dirname + '/lib';
+    config.entry = {
+        'webgl-operate': ['require.ts', 'polyfill.ts', 'webgl-operate.ts']
+    };
 
-module.exports.output.library = undefined;
-module.exports.output.libraryTarget = 'commonjs2';
+    config.module.rules[0].use.options.compilerOptions.noUnusedLocals = true;
+    config.module.rules[0].use.options.compilerOptions.declaration = true;
+    config.module.rules[0].use.options.compilerOptions.removeComments = false;
 
-module.exports.plugins[0].definitions = {
-    DISABLE_ASSERTIONS: JSON.stringify(false),
-    LOG_VERBOSITY_THRESHOLD: JSON.stringify(1),
+    config.output.library = undefined;
+    config.output.libraryTarget = 'commonjs2';
+
+    config.plugins[0].definitions = {
+        // DISABLE_ASSERTIONS: JSON.stringify(options.mode === 'development'),
+        DISABLE_ASSERTIONS: JSON.stringify(false),
+        // Log verbosity levels: debug = 3, info = 2, warn = 1, error = 0
+        LOG_VERBOSITY_THRESHOLD: JSON.stringify(options.mode === 'development' ? 3 : 2),
+    };
+
+    return config;
 };
